+++ src/Node/Comment.php
@@ -31,16 +31,6 @@ final class Comment implements Implementation
     }
 
     #[\Override]
-    public function toString(): string
-    {
-        $writer = new \XMLWriter;
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->openMemory();
-
-        return $this->render($writer);
-    }
-
-    #[\Override]
     public function render(\XMLWriter $writer): string
     {
         /** @psalm-suppress ImpureMethodCall */