+++ src/Node/Implementation.php
@@ -10,6 +10,5 @@ namespace Innmind\Xml\Node;
 interface Implementation
 {
     public function content(): string;
-    public function toString(): string;
     public function render(\XMLWriter $writer): string;
 }